Do not include ORCID icons in citations when exporting.
people_with_orcids compiled EPrints XML function used in citations should not include ORCIDs in exports.  This is because it relies on the ORCID icon being local or even if a full URL was used it would still require EPrints to be hit whenever the exported citations are loaded.  Also, the ORCID popup relies on CSS not available after export.  There may be better ways to provide this (e.g. icon included as base64 encoded URL of the image) but until then it is best to not include this in exports.